fix bugs with sorting by score

+++ Bot Application1/Intent.cs
@@ -9,7 +9,7 @@ namespace Bot_Application1
     public class Intent
     {
         public string intent { get; set; }
-        public double score { get; set; }
+        public decimal score { get; set; }
 
     }
 }
